// apps/api/src/proto/v1/booking/booking.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.6
//   protoc               unknown
// source: v1/booking/booking.proto

/* eslint-disable */
import type { Metadata } from "@grpc/grpc-js";
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";
import { Any } from "../../google/protobuf/any";

export const protobufPackage = "v1.booking";

/** BookingAppointment */
export interface BookingCreate {
  dateTime: string;
  duration: number;
  hostId: string;
  confirmationToken?: string | undefined;
  paymentMethod?: string | undefined;
  encAccessCode: Uint8Array;
  hostSessionKeys: BookingCreate_SessionKey[];
  guestSessionKeys: BookingCreate_SessionKey[];
  salt: Uint8Array;
}

export interface BookingCreate_SessionKey {
  credentialId: string;
  keyCiphertext: Uint8Array;
  keyType: string;
}

export interface BookingCreateResponse {
  paymentLink?: string | undefined;
  paymentIntent?: string | undefined;
  clientSecret?: string | undefined;
  bookingId?: string | undefined;
}

/** BookingResponse */
export interface BookingResponse {
  status: string;
  statusCode: number;
  error?: string | undefined;
  created?: BookingCreateResponse | undefined;
  listed?: BookingListResponse | undefined;
}

/** BookingList */
export interface BookingList {
  startDate?: string | undefined;
  endDate?: string | undefined;
  duration?: number | undefined;
}

export interface BookingListResponse {
  items?: Any | undefined;
}

export const V1_BOOKING_PACKAGE_NAME = "v1.booking";

/** Booking Service */

export interface BookingServiceClient {
  /** Book an appointment */

  create(request: BookingCreate, metadata: Metadata, ...rest: any): Observable<BookingResponse>;

  /** List */

  list(request: BookingList, metadata: Metadata, ...rest: any): Observable<BookingResponse>;
}

/** Booking Service */

export interface BookingServiceController {
  /** Book an appointment */

  create(
    request: BookingCreate,
    metadata: Metadata,
    ...rest: any
  ): Promise<BookingResponse> | Observable<BookingResponse> | BookingResponse;

  /** List */

  list(
    request: BookingList,
    metadata: Metadata,
    ...rest: any
  ): Promise<BookingResponse> | Observable<BookingResponse> | BookingResponse;
}

export function BookingServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["create", "list"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("BookingService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("BookingService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const BOOKING_SERVICE_NAME = "BookingService";